Floor Up and Down buttons
Added members to floor in order to process what the elevator should be 
doing when in idle mode.  I did not like resetting to the lobby... as it 
counted as a non-work move.

Set idle event to process if there were any up or down buttons pressed 
on the floors.